backend/requirements.txt
asgiref==3.5.0
django==3.2.13
pytz==2022.1
sqlparse==0.4.2
typing-extensions==4.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
django-extra-fields==3.0.2